Thread safe update to iToa and uToa

+++ ici/include/platform.h
@@ -1022,8 +1022,8 @@ extern void			killResourceLock(ResourceLock *);
 extern void			lockResource(ResourceLock *);
 extern void			unlockResource(ResourceLock *);
 
-extern char			*itoa(int);
-extern char			*utoa(unsigned int);
+extern char			*itoa(long long);
+extern char			*utoa(unsigned long long);
 #define postErrmsg(txt, arg)	_postErrmsg(__FILE__, __LINE__, txt, arg)
 extern void			_postErrmsg(const char *, int, const char *,
 					const char *);
